Move search_names() to middleware/indexing.py
